fix(exits): change value and revenue fields from int to float for accurate financial representation

+++ app/features/exits/models/exits_responses.py
@@ -5,7 +5,7 @@ from pydantic import BaseModel, Field
 class ExitResponse(BaseModel):
     id: int
     plate: str
-    value: int
+    value: float
     payment_method: str = "No registrado"
     created_at: datetime
 
@@ -15,7 +15,7 @@ class ExitStatsResponse(BaseModel):
     today_exits: int
     this_week_exits: int
     this_month_exits: int
-    total_revenue: int
-    today_revenue: int
-    this_week_revenue: int
-    this_month_revenue: int
+    total_revenue: float
+    today_revenue: float
+    this_week_revenue: float
+    this_month_revenue: float
